refactor: 결과 페이지 curation section 경로/컴포넌트명 정리
