+++ src/pages/generate/pages/result/curationSection/CardProductItem.tsx
