+++ src/pages/generate/pages/result/curationSection/CurationSection.css.ts
