+++ src/pages/generate/pages/result/curationSection/CurationSection.tsx
@@ -22,7 +22,7 @@ import { useSavedItemsStore } from '@/store/useSavedItemsStore';
 import { getGeneratedImageProducts } from '@pages/generate/apis/furniture';
 
 import CardProductItem from './CardProductItem';
-import * as styles from './CurationSheet.css';
+import * as styles from './CurationSection.css';
 
 import type { FurnitureProductsInfoResponse } from '@pages/generate/types/furniture';
 
@@ -72,7 +72,7 @@ const normalizeColorHexes = (value: unknown) => {
     : [...RESULT_CARD_UI_FALLBACK.colorHexes];
 };
 
-interface CurationSheetProps {
+interface CurationSectionProps {
   groupId?: number | null;
 }
 
@@ -81,7 +81,7 @@ interface CurationSheetProps {
  * - 감지된 가구 카테고리/상품을 고정 영역에 표시
  * - 그룹 기반 진입 시 groupId를 통해 캐시·프리패치 범위를 확정
  */
-export const CurationSheet = ({ groupId = null }: CurationSheetProps) => {
+export const CurationSection = ({ groupId = null }: CurationSectionProps) => {
   const activeImageId = useActiveImageId();
   const imageState = useActiveImageCurationState();
   const selectedCategoryId = imageState?.selectedCategoryId ?? null;
@@ -130,8 +130,12 @@ export const CurationSheet = ({ groupId = null }: CurationSheetProps) => {
         ? byProductId
         : index + 1;
 
-      const originalPrice = toFiniteNumber(product.furnitureProductOriginalPrice);
-      const discountPrice = toFiniteNumber(product.furnitureProductDiscountPrice);
+      const originalPrice = toFiniteNumber(
+        product.furnitureProductOriginalPrice
+      );
+      const discountPrice = toFiniteNumber(
+        product.furnitureProductDiscountPrice
+      );
       const discountRate = toFiniteNumber(product.furnitureProductDiscountRate);
       const saveCount = toFiniteNumber(product.furnitureProductSaveCount);
