+++ src/pages/generate/pages/result/curationSection/CurationSectionWrapper.css.ts
